Update serialisation.php - utf8_encode is deprecated in php8.2
In php8.2 utf8_encode nad utf8_decode is depricated.  You are already checking for mbstring extension and functions and if available you were then encoding the string.  This is simply setting the function to utilize the mbstring method to actually encode the string in a matter thatis compatible with php8.2.

https://php.watch/versions/8.2/utf8_encode-utf8_decode-deprecated#utf8_encode-any-mbstring